Fibonacci 50 correction 6

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,7 +9,7 @@
 int main(void)
 {
 	int i;
-	long int fib_sequence[100];
+	long long int fib_sequence[100];
 
 	fib_sequence[0] = 0;
 	fib_sequence[1] = 1;
